Add kitchen test for ssl_ciphers and ssl_versions
This came about because of:
https://github.com/rabbitmq/chef-cookbook/pull/501 and can be considered
a reference commit/branch. Not intending this gets merged in.

I was specifically interested in testing Chef 12.10.24 and Ubuntu 14.04
(and 16.04).

tl;dr: I cannot reproduce the issues we see in that PR 501 via Test
Kitchen. Which makes no sense, but there you go.

- Adds some self-signed certs to a `test/sslcert` folder
- Uses `synced_folders` in `.kitchen.yml` so we can use these certs
within the test kitchen VM
- Sets Chef version to 12.10.24
- Adds a Ubuntu 14.04 platform
- Copies the `default-deb` suite to `default-deb-ssl` and ...
    - Sets rabbitmq version to 3.7.9
    - Sets Erlang to 21.x
    - Sets ssl to true and references the self-signed certs
    - Sets ssl_ciphers and ssl_versions
- Adds a `test/smoke/ssl/default_spec.rb` file. This is pretty simple,
it just checks the `/etc/rabbitmq/rabbitmq.config` for the ssl ciphers
and versions being set as expected.

Note, in order to get this to work on Ubuntu 14.04 I had to edit the
Berksfile.lock and set:

- erlang to 6.2.0 (instead of 7.0.0)
- seven_zip to 2.0.2 (instead of 3.1.0); Even though irrelevant as
Windows

I couldn't find anyway of doing this apart from editing the
Berksfile.lock. I.e. just setting `recipe[erlang@6.2.0]` in the
`.kitchen.yml` was not enough (and probably irrelevant).

Anyway, I had hoped this would reproduce the issues we saw in our
production environment with Ubuntu 14.04 and Chef 12.10.24, but it
doesn't: This test passes.

There must be further differences. I think the Ruby version is probably
important. In our production environment we have 2.3.0 and this kitchen
test results in 2.1.0 - I don't know how to force a different version
via kitchen.

This test is run with: `rake integration:kitchen:default-deb-ssl-ubuntu-1404`